+++ cli/src/cleaner/for_tests/test5/data.json
@@ -1,0 +1,48 @@
+{
+  "polkadot": {
+    "title": "polkadot",
+    "color": "#9C9C9C",
+    "rpcEndpoint": "wss://rpc.polkadot.io",
+    "genesisHash": "0x91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3",
+    "unit": "DOT",
+    "base58prefix": 0,
+    "logo": "polkadot",
+    "decimals": 10,
+    "metadataQr": {
+      "version": 9,
+      "file": {
+        "path": "qr/polkadot_metadata_10.apng",
+        "signedBy": "Parity"
+      }
+    },
+    "liveMetaVersion": 10,
+    "latestMetadata": "",
+    "specsQr": {
+      "path": "/qr/polkadot_specs.png",
+      "signedBy": "Parity"
+    }
+  },
+  "polkadot-statemint": {
+    "title": "statemint",
+    "color": "#9C9C9C",
+    "rpcEndpoint": "wss://rpc.polkadot.io",
+    "genesisHash": "0x91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3",
+    "unit": "DOT",
+    "base58prefix": 0,
+    "logo": "polkadot",
+    "decimals": 10,
+    "metadataQr": {
+      "version": 9,
+      "file": {
+        "path": "qr/polkadot-statemint_metadata_9.apng",
+        "signedBy": "Parity"
+      }
+    },
+    "liveMetaVersion": 9,
+    "latestMetadata": "",
+    "specsQr": {
+      "path": "/qr/polkadot-statemint_specs.png",
+      "signedBy": "Parity"
+    }
+  }
+}
